Removing static site, keeping stats separately
