+++ markdown/six-principles/transcribed/book.toml
@@ -4,6 +4,7 @@ authors = ["William Perkins"]
 language = "en"
 multilingual = false
 src = "./"
+summary = "\"To bring you to true knowledge, vnfained faith, and ſound repentance: here I haue ſet downe the principall points of Chriſtian religion in ſixe plaine and eaſie rules: euen ſuch as the ſimpleſt may eaſily learne: and hereunto is adioyned and expoſition of them word by word.\" - William Perkins"
 
 [build]
 build-dir = "../../../dist/t/six-principles"
@@ -13,5 +14,5 @@ default-theme = "rust"
 preferred-dark-theme = "coal"
 git-repository-url = "https://github.com/glitchassassin/william-perkins"
 edit-url-template = "https://github.com/glitchassassin/william-perkins/edit/master/{path}"
-site-url = "/william-perkins/t/six-principles"
+site-url = "/t/six-principles"
 additional-css = ["themes/themes.css"]
