--- site/.gitignore
@@ -1,4 +1,0 @@
-dist
-node_modules
-src/_includes/css
-src/_data/transcriptions.json
