--- site/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "works-of-william-perkins",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "build:prebuild": "npm-run-all build:stats build:sass",
-    "build:stats": "node utils/calc-stats.js",
-    "build:eleventy": "eleventy",
-    "build:sass": "sass --style=compressed src/_scss/styles.scss src/_assets/css/styles.css",
-    "build": "npm-run-all build:prebuild build:eleventy",
-    "watch:eleventy": "eleventy --serve",
-    "watch:sass": "npm run build:sass -- --watch",
-    "start": "npm-run-all build:prebuild --parallel watch:*"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@11ty/eleventy": "^1.0.0",
-    "@fortawesome/fontawesome-free": "^6.1.1",
-    "npm-run-all": "^4.1.5",
-    "sass": "^1.50.0"
-  }
-}
