--- site/src/_assets/css/styles.css
@@ -1,1 +1,0 @@
-@import"https://fonts.googleapis.com/css2?family=Montserrat&Playfair+Display&family=Uncial+Antiqua&display=swap";html,body,div,span,applet,object,iframe,h1,h2,h3,h4,h5,h6,p,blockquote,pre,a,abbr,acronym,address,big,cite,code,del,dfn,em,img,ins,kbd,q,s,samp,small,strike,strong,sub,sup,tt,var,b,u,i,center,dl,dt,dd,ol,ul,li,fieldset,form,label,legend,table,caption,tbody,tfoot,thead,tr,th,td,article,aside,canvas,details,embed,figure,figcaption,footer,header,hgroup,menu,nav,output,ruby,section,summary,time,mark,audio,video{margin:0;padding:0;border:0;font-size:100%;font:inherit;vertical-align:baseline}article,aside,details,figcaption,figure,footer,header,hgroup,menu,nav,section{display:block}body{line-height:1}ol,ul{list-style:none}blockquote,q{quotes:none}blockquote:before,blockquote:after,q:before,q:after{content:"";content:none}table{border-collapse:collapse;border-spacing:0}.home h1,.home h2,.home h3{font-family:"Uncial Antiqua",cursive}.home .transcriptions{display:grid;grid-template-columns:repeat(auto-fit, minmax(250px, 1fr));gap:1em 1em;justify-items:center}h1,h2,h3{font-family:"Playfair Display",serif;text-align:center;font-weight:normal;line-height:1.2em}h1{font-size:4em}h2{font-size:2em}h3{font-size:1.5em}body{font-family:"Montserrat",serif;font-size:1.2em}p{margin:2em;line-height:1.2em}.text-center{text-align:center}a{color:#fff}ul{list-style:disc;padding-left:1em;margin:2em}ol{list-style:decimal;padding-left:1em;margin:2em}@media(min-width: 768px){main>ul,main>ol{margin:2em 4em}}li{line-height:1.2em}blockquote{margin-bottom:1em;border-left:3px solid #fff;padding:1em;padding-left:2em;font-style:italic;background-color:rgba(255,255,255,.0666666667)}@media(min-width: 768px){blockquote{margin-left:2em}}blockquote p{margin:0}em{font-style:italic}strong{font-weight:bold}.stage>header ul{height:100%;list-style-type:none;display:grid;grid-template-columns:1fr min-content;justify-content:end;align-items:center;gap:1em;margin:1em}.stage>header i{color:#fff;font-size:2em}.stage>header .logo-font{font-family:"Uncial Antiqua";justify-self:start;font-size:1.5em}.stage>header .logo-font a{text-decoration:none}.stage>footer{height:75px;display:grid;justify-items:end;align-items:center}.stage>footer .credits{color:#fff;text-align:right;font-size:.7em}.stage>footer .credits a{color:#fff}.transcription-progress{background-color:rgba(255,255,255,.0666666667);padding-top:.5em;height:300px;width:100%;max-width:300px;color:inherit;text-decoration:none;display:grid;grid-template-rows:100px 1fr 25px;align-items:end}.transcription-progress:hover{background-color:rgba(255,255,255,.1333333333)}.transcription-progress>.status{text-align:center;padding:1em}.transcription-progress>.progress{box-shadow:0 5px 5px -5px #999 inset;background-color:rgba(255,255,255,.0666666667);height:25px}.transcription-progress>.progress>span{height:inherit;box-shadow:0 5px 5px -5px #999 inset;background-color:#393;display:block;white-space:nowrap}.volumes-list{display:grid;grid-template-columns:repeat(auto-fit, minmax(250px, 1fr));gap:1em 1em;justify-items:center;list-style:none;margin:1em 0;padding:0}.volumes-list a{background-color:rgba(255,255,255,.0666666667);width:100%;color:inherit;text-decoration:none;display:inline-block}.volumes-list a:hover{background-color:rgba(255,255,255,.1333333333)}.volumes-list a>h3{padding:1em}body{background-color:#222;color:#fff;margin:0;height:100vh;background-image:url("../christs-college.jpg");background-size:cover;background-position:top center;background-repeat:no-repeat;background-attachment:fixed}.stage{max-width:1200px;background-color:rgba(34,34,34,.9333333333);min-height:100%;margin:0 auto;padding:0 20px;display:grid;grid-template-rows:min-content 1fr min-content}/*# sourceMappingURL=styles.css.map */
