--- site/src/_github/CNAME
@@ -1,1 +1,0 @@
-perkins.kingsquill.org
