--- site/src/_includes/footer.njk
@@ -1,5 +1,0 @@
-<footer>
-  <div class="credits">
-    Christ's College: <a href="http://www.royan.com.ar">© Jorge Royan / CC BY-SA 3.0</a>
-  </div>
-</footer>
