--- site/src/_includes/header.njk
@@ -1,19 +1,0 @@
-<header class="site-header">
-  <nav>
-    <ul>
-      <li class="logo-font">
-        <a href="{{'/' | url}}" title="Home">
-          WP
-        </a>
-      </li>
-      {# <li>
-        <a href='https://ko-fi.com/T6T6C4TMR' target='_blank'><img height='36' style='border:0px;height:36px;' src='https://cdn.ko-fi.com/cdn/kofi2.png?v=3' border='0' alt='Buy Me a Coffee at ko-fi.com'/></a>
-      </li> #}
-      <li>
-        <a href="https://github.com/glitchassassin/william-perkins" title="Contribute at GitHub">
-          <i class="fa-brands fa-github"></i>
-        </a>
-      </li>
-    </ul>
-  </nav>
-</header>
