--- site/src/_scss/footer.scss
@@ -1,16 +1,0 @@
-.stage>footer {
-  height: 75px;
-  display: grid;
-  justify-items: end;
-  align-items: center;
-
-  .credits {
-    color: #fff;
-    text-align: right;
-    font-size: 0.7em;
-
-    a {
-      color: #fff;
-    }
-  }
-}
