--- site/src/_scss/header.scss
@@ -1,27 +1,0 @@
-.stage>header {
-    ul {
-        height: 100%;
-        list-style-type: none;
-        display: grid;
-        grid-template-columns: 1fr min-content;
-        justify-content: end;
-        align-items: center;
-        gap: 1em;
-        margin: 1em;
-    }
-
-    i {
-        color: #fff;
-        font-size: 2em;
-    }
-
-    .logo-font {
-        font-family: 'Uncial Antiqua';
-        justify-self: start;
-        font-size: 1.5em;
-
-        a {
-            text-decoration: none;
-        }
-    }
-}
