--- site/src/_scss/pages/home.scss
@@ -1,15 +1,0 @@
-.home {
-
-    h1,
-    h2,
-    h3 {
-        font-family: 'Uncial Antiqua', cursive;
-    }
-
-    .transcriptions {
-        display: grid;
-        grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-        gap: 1em 1em;
-        justify-items: center;
-    }
-}
