--- site/src/_scss/styles.scss
@@ -1,29 +1,0 @@
-@import './reset.scss';
-@import './pages/home.scss';
-@import './typography.scss';
-@import './header.scss';
-@import './footer.scss';
-@import './widgets.scss';
-
-body {
-    background-color: #222;
-    color: #fff;
-    margin: 0;
-    height: 100vh;
-    background-image: url('../christs-college.jpg');
-    background-size: cover;
-    background-position: top center;
-    background-repeat: no-repeat;
-    background-attachment: fixed;
-}
-
-.stage {
-    max-width: 1200px;
-    background-color: #222222ee;
-    min-height: 100%;
-    margin: 0 auto;
-    padding: 0 20px;
-
-    display: grid;
-    grid-template-rows: min-content 1fr min-content;
-}
