--- site/src/_scss/typography.scss
@@ -1,91 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Montserrat&Playfair+Display&family=Uncial+Antiqua&display=swap');
-
-h1,
-h2,
-h3 {
-    font-family: 'Playfair Display', serif;
-    text-align: center;
-    font-weight: normal;
-    line-height: 1.2em;
-}
-
-h1 {
-    font-size: 4em;
-}
-
-h2 {
-    font-size: 2em;
-}
-
-h3 {
-    font-size: 1.5em;
-}
-
-body {
-    font-family: 'Montserrat', serif;
-    font-size: 1.2em;
-}
-
-p {
-    margin: 2em;
-    line-height: 1.2em;
-}
-
-.text-center {
-    text-align: center;
-}
-
-a {
-    color: #fff;
-}
-
-ul {
-    list-style: disc;
-    padding-left: 1em;
-    margin: 2em;
-}
-
-ol {
-    list-style: decimal;
-    padding-left: 1em;
-    margin: 2em;
-}
-
-@media (min-width: 768px) {
-
-    main>ul,
-    main>ol {
-        margin: 2em 4em;
-    }
-}
-
-li {
-    line-height: 1.2em;
-}
-
-blockquote {
-    @media (min-width: 768px) {
-        margin-left: 2em;
-    }
-
-    margin-bottom: 1em;
-    border-left: 3px solid #fff;
-    padding: 1em;
-    padding-left: 2em;
-
-    font-style: italic;
-
-    background-color: #ffffff11;
-
-    p {
-        margin: 0;
-    }
-}
-
-em {
-    font-style: italic;
-}
-
-strong {
-    font-weight: bold;
-}
