--- site/src/_scss/widgets.scss
@@ -1,65 +1,0 @@
-.transcription-progress {
-  &:hover {
-    background-color: #ffffff22;
-  }
-
-  background-color: #ffffff11;
-  padding-top: 0.5em;
-  height: 300px;
-  width: 100%;
-  max-width: 300px;
-
-  color: inherit;
-  text-decoration: none;
-
-  display: grid;
-  grid-template-rows: 100px 1fr 25px;
-  align-items: end;
-
-  >.status {
-    text-align: center;
-    padding: 1em;
-  }
-
-  >.progress {
-    box-shadow: 0 5px 5px -5px #999 inset;
-    background-color: #ffffff11;
-    height: 25px;
-
-    >span {
-      height: inherit;
-      box-shadow: 0 5px 5px -5px #999 inset;
-      background-color: #393;
-      display: block;
-      white-space: nowrap;
-    }
-  }
-}
-
-.volumes-list {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-  gap: 1em 1em;
-  justify-items: center;
-  list-style: none;
-  margin: 1em 0;
-  padding: 0;
-
-  a {
-    &:hover {
-      background-color: #ffffff22;
-    }
-
-    background-color: #ffffff11;
-    width: 100%;
-
-    color: inherit;
-    text-decoration: none;
-
-    display: inline-block;
-
-    >h3 {
-      padding: 1em;
-    }
-  }
-}
