--- site/src/index.njk
@@ -1,27 +1,0 @@
----
-layout: layouts/default
-tags: home
----
-
-{% from "widgets/transcription-progress.njk" import transcriptionProgress %}
-
-<h1>The Works</h1>
-<h2>of William Perkins</h2>
-
-<p class="text-center">Transcribed by volunteers and freely distributed in the public domain.</p>
-
-<ul class="volumes-list">
-{%- for volume in collections.volumes -%}
-  <li><a href="{{volume.url | url}}"><h3>{{ volume.data.title }}</h3></a></li>
-{%- endfor -%}
-</ul>
-
-<p class="text-center">Books in Progress</p>
-
-<div class="transcriptions">
-  {% for transcription in transcriptions %}
-  {% if transcription.progress < 100 %}
-    {{ transcriptionProgress(transcription.name, transcription.link, transcription.status, transcription.progress) }}
-  {% endif %}
-  {% endfor %}
-</div>
