--- site/src/volume/1/index.md
@@ -1,29 +1,0 @@
----
-layout: layouts/default
-tags: volumes
-title: Volume 1
----
-
-{% include "./title.md" %}
-
-## Table of Contents
-
-- [The Printer to the Reader](printers-preface)
-- [Six Principles: The Foundation of Christian Religion](../../six-principles/) ([original transcription](../../t/six-principles/))
-- [A Golden Chain: The Description of Theology](../../golden-chain) ([original transcription](../../t/golden-chain))
-- *An Exposition of the Creed of the Apostles*
-- *An Exposition of the Lord's Prayer*
-- *Damnation or Grace*
-- *A Case of Conscience*
-- *The Government of the Tongue*
-- *Two Treatises*
-- *On Living Well*
-- *On Dying Well*
-- *On Conscience*
-- *A Reformed Catholic*
-- *Knowing Christ Crucified*
-- *A Grain of Mustard Seed*
-- *The True Gain*
-- *A Warning Against Idolatry*
-- *On God's Free Grace and Man's Free Will*
-- *On Callings*
