--- site/src/volume/1/printers-preface.md
@@ -1,23 +1,0 @@
----
-layout: layouts/default
-title: The Printer to the Reader.
----
-
-# Volume 1
-
-## The Printer to the Reader.
-
-I suppose it altogether needless (Christian Reader) by commending Mr. William Perkins, the Author of this Book, to woo your holy affection, which either himself in his life time by his Christian conversation hath wooed in you, or since his death, the never-dying memory of his excellent knowledge, his great humility, his sound religion, his fervent zeal, his painful labours, in the Church of God, do most justly challenge at your hands. Only in one word I dare be bold to say of him, as in times past Nazianzen spake of Athanasius: "His life was a good definition of a true Minister and preacher of the Gospel." And needless also is it (if not injurious) by praising of his learned and godly writings, to think to incline your better judgement which in so holy a subject (as is the body of Divinity, and principles of Religion) I dare not suspect to be unsettled; and in so easy a method and familiar a style, which he useth in them both, I am sure can hardly miscarry. Such is his love to all, that, as for the matter of his doctrine, he contenteth and satisfieth the most learned; so for the manner of his delivering the same, he condescendeth to the capacity of the meanest of God's children. I here once more present unto you the works of this worthy man. Read them diligently, and judge of them freely. I doubt not but in your exactest censure, you will conspire with those learned Men, who for the profitable instruction they contain in all, or the most points of Christianity, for the more common good of the Church of God, have deemed them worthy their godly labours, by translating them into divers languages, as into Latin, Dutch, Spanish, &c. A thing not ordinarily observed in other writings of these our times.
-
-Concerning this new Edition, if you ask, why I have not added these other Treatises of his, which have been published since the former Impression? I answer, it is not done without the advice of grave and learned men in two respects: First, to set a difference (as indeed there is great odds) betwixt those books which the Author himself yet living, by his own care not only in the penning, but likewise in the correcting of the same, did set forth: and those other which are *Posthumi,* that is, born after his death, gathered and collected by others: Such as are these following:
-
-1. Upon three chapters of the Revelation.
-2. Satan's Sophistry.
-3. Upon the Epistle of Saint Jude.
-4. The dignity of the Ministry.
-5. Upon the eleventh chapter to the Hebrews.
-6. Upon three chapters of Saint Matthew's Gospel, &c.
-
-Secondly, because I have remaining by me so many other Treatises in divers prints and forms of the same Author's own collection, which I purpose (by God's permission) to set forth, as will amount to a second Volume equal to this, if not bigger.
-
-In the mean time, what heretofore I have wished concerning such as have the interest in those Afterbirths, *viz.* that they would agree likewise, to draw them all into one Volume, that so they that take delight in Mr. Perkins' writings, might with less labour of their own, attain to the complete body of his studies: it being also more convenient for their use, to have them rather compacted in one Volume, than scattered into divers parcels of divers forms: I now understand, is effected, and nothing henceforth to be added. Thus with promise to make all possible expedition in setting forth the second Volume of his works: I commend this to your diligent reading, my endeavours to your charitable construction, and us all to the blessing and holy protection of Almighty God. *From London this 15th of December, 1612.*
