--- site/src/volume/1/title.md
@@ -1,17 +1,0 @@
-# THE WORKS
-
-## OF THAT FAMOUS AND WORTHY MINISTER OF CHRIST in the University of Cambridge, MR. WILLIAM PERKINS.
-
-### THE FIRST VOLUME: NEWLY CORRECTED ACCORDING TO HIS OWN COPIES.
-
-With distinct chapters, and contents of every Book, and two Tables of the whole: one, of the matter and questions, the other of choice places of Scripture.
-
-> Isaiah 57:1  
-> The righteous perisheth, and no man considereth it in heart: merciful men are taken away, and no man understandeth that the righteous is taken away from the evil to come.
-
-> 2 Peter 1:15  
-> Nevertheless, I will endeavour always, that ye also may be able to have remembrance of these things after my departing.
-
-London,  
-Printed by **John Legatt.**  
-1626.
