+++ stats/README.md
@@ -1,0 +1,3 @@
+# Website Build
+
+The front page of the website is built with [Eleventy](https://www.11ty.dev/).
